Add support for changing the default git-host and defaulting the prefix
Up until now a package-source such as `jan9103/packer.nu`
always pointed to github. With this change you can configure
which git-host is used in this case.
You can configure it in your packages.nuon with the `default_git_host`
variable.

This also adds support for defaulting the git-repo prefix (author /
organisation / …) with the `default_git_repo_prefix` `packages.nuon`
variable. If the variable for example is set to `packer.nu` `foo` is
expanded to `packer.nu/foo`.