Update install script and remove broken bootstrap
Nushell first checks if a entire script is ready and then runs it,
which usually is a good thing, but not if i want to first create a
file (during bootstrap) and then use it afterwards.
I will NOT move the `overlay use packer_api` to the `config` since
if i do that a single broken packer-package would not just prevent
all other packages from loading, but the, which is used to fix the
issue as well.